requirements-telegram-bot.txt
aiogram>=3.13.0
tenacity>=8.0.0
httpx>=0.27.0
openpyxl>=3.1.0
PyYAML>=6.0
python-dotenv>=1.0.0
